several int to size_t changes based on compiler warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@101950 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objmgr/util/indexer.cpp
@@ -3695,7 +3695,7 @@ static const char* const mixedTags[] = {
 static int SkipMixedContent ( const char* ptr )
 
 {
-    for (int i = 0; i < sizeof (mixedTags); i++) {
+    for (size_t i = 0; i < sizeof (mixedTags); i++) {
         const char* tag = mixedTags[i];
         const char* tmp = ptr;
         int len = 0;
@@ -3772,7 +3772,7 @@ void CWordPairIndexer::PopulateWordPairIndex (string str)
     }
 
     // convert non-alphanumeric punctuation to space
-    for (int i = 0; i < str.length(); i++) {
+    for (size_t i = 0; i < str.length(); i++) {
         char ch = str[i];
         if (ch >= 'A' && ch <= 'Z') {
         } else if (ch >= 'a' && ch <= 'z') {
